Move stringifyException from VayraMergedModPlugin to it's proper new place
